+++ cuda/octree/src/cuda/bfrs.cu
@@ -96,6 +96,6 @@ void pcl::device::bruteForceRadiusSearch(const OctreeImpl::PointCloud& cloud, co
     
     //main bottle neck is a kernel call overhead/allocs
     //work time for 871k points ~0.8ms
-    int count = thrust::copy_if(first, last, cloud_ptr, res_ptr, cond) - res_ptr;
+    int count = (int)(thrust::copy_if(first, last, cloud_ptr, res_ptr, cond) - res_ptr);
     result = DeviceArray<int>(buffer.ptr(), count);
 }
